+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
 
+### Added
+- Added `sts-generate-voice-casting-prompt-file` utility to aid in casting voice configs
+
 ## [1.1.0] - 2025-05-26
 
 ### Added